+++ mods/include/mods/shading/shading.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void safe_normal(float x, float y, float z);