+++ mods/src/misc/internal.h
@@ -5,6 +5,7 @@ MCPI_INTERNAL void _init_misc_api();
 MCPI_INTERNAL void _init_misc_graphics();
 MCPI_INTERNAL void _init_misc_ui();
 MCPI_INTERNAL void _init_misc_home();
+MCPI_INTERNAL void _init_misc_item_rendering();
 
 template <typename... Args>
 static void nop(MCPI_UNUSED Args... args) {