+++ mods/src/misc/misc.cpp
@@ -688,4 +688,5 @@ void init_misc() {
     _init_misc_api();
     _init_misc_graphics();
     _init_misc_home();
+    _init_misc_item_rendering();
 }